bot/requirements.txt
torch==1.13.1
transformers==4.21.0
python-telegram-bot==21.6
python-dotenv==1.0.1